+++ src/main/PAL/lib/libsnd/SsStartSoundTickMode1.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-void SsStartSoundTick(s32 arg0) asm("func_80071C44");
+void SsStartSoundTick(long arg0) asm("func_80071C44");
 
 void SsStartSoundTickMode1(void) asm("func_80071EAC");
