+++ src/main/PAL/lib/libsnd/SsUtSetReverbType.c
@@ -2,23 +2,23 @@
 #include "psyq/spu.h"
 
 extern SpuReverbAttr D_8019C778;
-extern s16 D_8019C77C;
+extern short D_8019C77C;
 
-s32 SsUtSetReverbType(s32 type) asm("func_80073614");
-s32 SsUtSetReverbType(s32 type) {
-    register s32 normalized asm("v1");
-    register s32 negative asm("a1");
-    register s32 mode asm("v0");
-    register s32 result asm("s0");
+long SsUtSetReverbType(long type) asm("func_80073614");
+long SsUtSetReverbType(long type) {
+    register long normalized asm("v1");
+    register long negative asm("a1");
+    register long mode asm("v0");
+    register long result asm("s0");
 
     negative = 0;
     normalized = type;
-    if ((s32)(type << 16) < 0) {
+    if ((long)(type << 16) < 0) {
         negative = 1;
         normalized = -type;
     }
 
-    if ((u32)(normalized & 0xFFFF) >= 10) {
+    if ((u_long)(normalized & 0xFFFF) >= 10) {
         goto fail;
     }
 
@@ -29,7 +29,7 @@ s32 SsUtSetReverbType(s32 type) {
         mode = normalized << 16;
     }
     mode >>= 16;
-    *(s32 *)&D_8019C77C = mode;
+    *(long *)&D_8019C77C = mode;
     mode = normalized << 16;
     result = mode >> 16;
 
@@ -47,8 +47,8 @@ fail:
     return -1;
 }
 
-s32 SsUtGetReverbType(void) asm("func_800736B8");
-s32 SsUtGetReverbType(void) {
+long SsUtGetReverbType(void) asm("func_800736B8");
+long SsUtGetReverbType(void) {
     return D_8019C77C;
 }
 
@@ -63,24 +63,24 @@ void SsUtReverbOff(void) {
 }
 
 
-void SsUtSetReverbFeedback(s32 feedback) asm("func_80073708");
-void SsUtSetReverbFeedback(s32 feedback) {
+void SsUtSetReverbFeedback(long feedback) asm("func_80073708");
+void SsUtSetReverbFeedback(long feedback) {
     D_8019C778.mask = 0x10;
-    D_8019C778.feedback = (s16)feedback;
+    D_8019C778.feedback = (short)feedback;
     SpuSetReverbModeParam(&D_8019C778);
 }
 
-void SsUtSetReverbDepth(s32 left, s32 right) asm("func_80073748");
-void SsUtSetReverbDepth(s32 left, s32 right) {
+void SsUtSetReverbDepth(long left, long right) asm("func_80073748");
+void SsUtSetReverbDepth(long left, long right) {
     D_8019C778.mask = 0x6;
-    D_8019C778.depth.left = ((s16)left * 0x7FFF) / 0x7F;
-    D_8019C778.depth.right = ((s16)right * 0x7FFF) / 0x7F;
+    D_8019C778.depth.left = ((short)left * 0x7FFF) / 0x7F;
+    D_8019C778.depth.right = ((short)right * 0x7FFF) / 0x7F;
     SpuSetReverbModeParam(&D_8019C778);
 }
 
-void SsUtSetReverbDelay(s32 delay) asm("func_800737E0");
-void SsUtSetReverbDelay(s32 delay) {
+void SsUtSetReverbDelay(long delay) asm("func_800737E0");
+void SsUtSetReverbDelay(long delay) {
     D_8019C778.mask = 0x8;
-    D_8019C778.delay = (s16)delay;
+    D_8019C778.delay = (short)delay;
     SpuSetReverbModeParam(&D_8019C778);
 }
