+++ src/main/PAL/lib/libsnd/_SsSndTempo.c
@@ -1,15 +1,15 @@
 #include "psyq/snd.h"
 
 extern SeqStruct *D_801E79CC[];
-extern s32 D_801E6C6C;
+extern long D_801E6C6C;
 
-void _SsSndTempo(s16 arg0, s16 arg1) {
+void _SsSndTempo(short arg0, short arg1) {
     SeqStruct *pSeq = &D_801E79CC[arg0][arg1];
 
     pSeq->tempo_countdown--;
 
     if (pSeq->tempo_step > 0) {
-        if ((pSeq->tempo_countdown % (u32)pSeq->tempo_step) == 0) {
+        if ((pSeq->tempo_countdown % (u_long)pSeq->tempo_step) == 0) {
             if (pSeq->tempo > pSeq->target_tempo) {
                 pSeq->tempo--;
             } else if (pSeq->tempo < pSeq->target_tempo) {
@@ -17,7 +17,7 @@ void _SsSndTempo(s16 arg0, s16 arg1) {
             }
 
             pSeq->tick_period =
-                (u32)((pSeq->tempo_multiplier * pSeq->tempo) * 0xA) / (u32)(D_801E6C6C * 0x3C);
+                (u_long)((pSeq->tempo_multiplier * pSeq->tempo) * 0xA) / (u_long)(D_801E6C6C * 0x3C);
             if (pSeq->tick_period <= 0) {
                 pSeq->tick_period = 1;
             }
@@ -40,7 +40,7 @@ void _SsSndTempo(s16 arg0, s16 arg1) {
         }
 
         pSeq->tick_period =
-            (u32)((pSeq->tempo_multiplier * pSeq->tempo) * 0xA) / (u32)(D_801E6C6C * 0x3C);
+            (u_long)((pSeq->tempo_multiplier * pSeq->tempo) * 0xA) / (u_long)(D_801E6C6C * 0x3C);
         if (pSeq->tick_period <= 0) {
             pSeq->tick_period = 1;
         }
