+++ src/main/PAL/lib/libspu/SpuSetNoiseVoice.c
@@ -1,5 +1,5 @@
 #include "psyq/spu.h"
 
-u32 SpuSetNoiseVoice(s32 on_off, u32 voice_bit) {
+u_long SpuSetNoiseVoice(long on_off, u_long voice_bit) {
     return _SpuSetAnyVoice(on_off, voice_bit, 0xCA, 0xCB);
 }
