+++ src/main/PAL/main/sdk/BiosFormatDevice.c
@@ -1,6 +1,8 @@
+#include <sys/types.h>
+
 #include "common.h"
 
-u32 func_80063280[4] __attribute__((section(".text"))) = {
+u_long func_80063280[4] __attribute__((section(".text"))) = {
     0x240A00B0,
     0x01400008,
     0x24090041,
